feat: add helper for test
so jest mocks will not be transferred from one test to the other
