+++ src/models/wasteWater/story/__tests__/WasteWaterSamplingSitesHooks.test.ts
@@ -1,13 +1,8 @@
-import {
-  filterByDateRange,
-  getMaxDateRange,
-  useWasteWaterData,
-  WasteWaterDataWithLocation,
-} from '../WasteWaterSamplingSitesHooks';
+import { filterByDateRange, getMaxDateRange, useWasteWaterData } from '../WasteWaterSamplingSitesHooks';
 import { getData } from '../../loading';
 import { renderHook, waitFor } from '@testing-library/react';
-import { WasteWaterTimeEntry } from '../../types';
 import { globalDateCache } from '../../../../helpers/date-cache';
+import { getTestWasteWaterDataWithLocation } from '../testHelper';
 
 jest.mock('../../loading');
 const getDataMock = getData as jest.Mock;
@@ -70,34 +65,6 @@ describe('filterByDateRange', function () {
     expect(filteredData[0].variantsTimeseriesSummaries[0].data.length).toEqual(2);
   });
 });
-
-export function getTestWasteWaterDataWithLocation(
-  dates = ['2021-01-01', '2021-01-02', '2021-01-03', '2021-01-04'],
-  variants = ['variantName1'],
-  locations = ['location1']
-): WasteWaterDataWithLocation[] {
-  const unifiedDays = dates.map(date => {
-    return globalDateCache.getDay(date);
-  });
-
-  return locations.map(location => {
-    return {
-      location: location,
-      variantsTimeseriesSummaries: variants.map(variant => {
-        return {
-          name: variant,
-          data: unifiedDays.map(date => {
-            return {
-              date,
-              proportion: 0.1,
-              proportionCI: [0.1, 0.1],
-            } as WasteWaterTimeEntry;
-          }),
-        };
-      }),
-    };
-  });
-}
 
 describe('getMaxDateRange', function () {
   it('should return minimum and maximum date of wasteWaterData', function () {
